unify module.c files from coord so that "onload" is on the module not coord
