+++ src/module-init/module-init.c
@@ -1,5 +1,4 @@
 
-#define REDISMODULE_MAIN
 #include "redismodule.h"
 
 #include "module.h"
